Clean up CFBundleDocumentTypes
Apple has the concept of Uniform Type Identifiers[1] which allows to
associate an app to a bunch of file-types at once. This commit add the
`public.text` TUI to the Info.plist file and removes unneeded
declarations. I didn't test all of them, but they can be added as needed

[1]: https://developer.apple.com/library/archive/documentation/FileManagement/Conceptual/understanding_utis