Fix for bad Windows shell integration
the setup process doesn't integrate with the shell properly. 

trying to open with the context menu gives this error dialog
![image](https://user-images.githubusercontent.com/55286472/174404139-952199e2-c66c-4e73-91b6-af684a3a1698.png)

i found the problem, after a bit. 

the registry value it gives ends up having a slash before the filepath to fvim 
so it would be something like 
"\C:\Fvim\Fvim.exe" "%1" 
instead of 
"C:\Fvim\Fvim.exe" "%1"

changing to the second version in my registry editor does actually fix the problem.